Refactor some files
